fix: update import path after restructuring of the packages
