+++ propra_webscience_ws24/training/classical/model/model_factory.py
@@ -1,15 +1,20 @@
 import pandas as pd
 
-from propra_webscience_ws24.training.model import RandomForest
-from propra_webscience_ws24.training.model.model_base import ModelType, ModelBase
-from propra_webscience_ws24.training.model.svm import SVM
-from propra_webscience_ws24.training.model.decision_tree import DecisionTree
-from propra_webscience_ws24.training.model.knn import KNN
-from propra_webscience_ws24.training.model.naive_bayes import NaiveBayes
-from propra_webscience_ws24.training.model.logistic_regression import (
+from propra_webscience_ws24.training.classical.model import RandomForest
+from propra_webscience_ws24.training.classical.model.model_base import (
+    ModelType,
+    ModelBase,
+)
+from propra_webscience_ws24.training.classical.model.svm import SVM
+from propra_webscience_ws24.training.classical.model.decision_tree import DecisionTree
+from propra_webscience_ws24.training.classical.model.knn import KNN
+from propra_webscience_ws24.training.classical.model.naive_bayes import NaiveBayes
+from propra_webscience_ws24.training.classical.model.logistic_regression import (
     LogisticRegressionModel,
 )
-from propra_webscience_ws24.training.training_combinations import TrainingCombination
+from propra_webscience_ws24.training.classical.training_combinations import (
+    TrainingCombination,
+)
 
 
 models = {
@@ -30,5 +35,5 @@ def model_factory(
     model_args: dict,
 ) -> ModelBase:
     if model_type in models:
-        return models[model_type](df_train, df_test, training_combination, model_args)
+        return models[model_type](df_train, df_test, training_combination, model_args)  # type: ignore[operator]
     raise ValueError(f"Unsupported model type: {model_type}")
