+++ propra_webscience_ws24/training/classical/model/naive_bayes.py
@@ -1,6 +1,9 @@
 from sklearn.naive_bayes import BernoulliNB
 
-from propra_webscience_ws24.training.model.model_base import ModelBase, ModelType
+from propra_webscience_ws24.training.classical.model.model_base import (
+    ModelBase,
+    ModelType,
+)
 
 
 class NaiveBayes(ModelBase[BernoulliNB]):
